Update to version 1.2
git-svn-id: https://plugins.svn.wordpress.org/delete-me/trunk@664738 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ fcn/deactivate.php
@@ -9,19 +9,56 @@ if ( isset( $this ) == false || get_class( $this ) != 'plugin_delete_me' ) {
 
 // Deactivate
 
-if ( $this->option['settings']['uninstall_on_deactivate'] == true ) {
+if ( $network_wide ) {
 	
-	// Remove capability
+	// Network deactivation
 	
-	foreach ( $this->wp_roles->role_objects as $role ) {
+	$blog_ids = $this->wpdb->get_col( $this->wpdb->prepare( "SELECT `blog_id` FROM " . $this->wpdb->blogs ) );
+	
+	foreach ( $blog_ids as $blog_id ) {
 		
-		$role->remove_cap( $this->info['cap'] );
+		switch_to_blog( $blog_id );
+		
+		$this->option = $this->fetch_option();
+		
+		if ( ! empty( $this->option ) && $this->option['settings']['uninstall_on_deactivate'] == true ) {
+			
+			// Remove capability
+			
+			foreach ( $this->wp_roles->role_objects as $role ) {
+				
+				$role->remove_cap( $this->info['cap'] );
+				
+			}
+			
+			// Delete option
+			
+			delete_option( $this->info['option'] );
+			
+		}
+		
+		restore_current_blog();
 		
 	}
 	
-	// Delete option
+} else {
 	
-	delete_option( $this->info['option'] );
+	// Site deactivation
+	
+	if ( $this->option['settings']['uninstall_on_deactivate'] == true ) {
+		
+		// Remove capability
+		
+		foreach ( $this->wp_roles->role_objects as $role ) {
+			
+			$role->remove_cap( $this->info['cap'] );
+			
+		}
+		
+		// Delete option
+		
+		delete_option( $this->info['option'] );
+		
+	}
 	
 }
-?>
